requirements.txt
fastapi
uvicorn
streamlit
keras
numpy
opencv-python
Pillow
tensorflow
python-multipart
streamlit-cropper
streamlit-authenticator
torch
torchvision
timm
